requirements.txt
Flask
Flask-SocketIO
websockets
cryptography
asgiref
PyNaCl

